pyproject.toml
[project]
name = "mail-digest-aws"
version = "0.1.0"
description = ""
authors = [{ name = "Matheus Santos", email = "matheusgabriel.18@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-openai (>=0.3.19,<0.4.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
]

[tool.poetry]
package-mode = false


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-cov = "^6.1.1"
taskipy = "^1.14.1"
ruff = "^0.11.13"
pytest-asyncio = "^1.0.0"
factory-boy = "^3.3.3"
freezegun = "^1.5.2"

[tool.ruff]
line-length = 79
exclude = ["tests/"]

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
pre_test = 'task lint'
test = 'pytest -s -x --cov=app -vv'
post_test = 'coverage html'

[tool.coverage.run]
concurrency = ["thread", "greenlet"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
